+++ src/libtrx/include/libtrx/game/ui/dialogs/new_game.h
@@ -5,12 +5,10 @@
 #include "../common.h"
 #include "../elements/requester.h"
 
-typedef struct {
-    UI_REQUESTER_STATE req;
-} UI_NEW_GAME_STATE;
+typedef struct UI_NEW_GAME_STATE UI_NEW_GAME_STATE;
 
 // state functions
-void UI_NewGame_Init(UI_NEW_GAME_STATE *s);
+UI_NEW_GAME_STATE *UI_NewGame_Init(void);
 int32_t UI_NewGame_Control(UI_NEW_GAME_STATE *s);
 void UI_NewGame_Free(UI_NEW_GAME_STATE *s);